backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-phonenumber-field==7.2.0
django-seed==0.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
Faker==20.0.3
fonttools==4.43.1
h11==0.14.0
idna==3.4
importlib-resources==6.1.0
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.2
phonenumbers==8.13.25
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
toposort==1.10
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
zipp==3.17.0